+++ plone/app/users/schema.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from Products.CMFCore.interfaces import ISiteRoot
 from Products.CMFCore.utils import getToolByName
-from Products.CMFDefault.exceptions import EmailAddressInvalid
+from Products.CMFPlone.RegistrationTool import EmailAddressInvalid
 from Products.CMFPlone import PloneMessageFactory as _
 from zope import schema
 from zope.component import getUtility